backups/backup_20251022_114036/requirements.txt
Django==5.1.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
requests==2.32.3
dj-database-url==2.2.0

# LLM SDKs
openai==1.40.0
google-generativeai==0.7.2
anthropic==0.34.0
huggingface-hub==0.24.5

# Additional utilities
Pillow==10.4.0
pyotp==2.9.0
qrcode==7.4.2

